VMS/VMS/requirements.txt
Django==4.2.3
django_jsonfield==1.4.1
djangorestframework==3.14.0
